--- Lesson20-UI-Reaction/src/main/java/DeleteServlet.java
@@ -1,23 +1,0 @@
-import domain.Fridge;
-
-import javax.servlet.ServletException;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import java.util.List;
-import java.util.UUID;
-
-@WebServlet("/delete")
-public class DeleteServlet extends HttpServlet {
-    private final CustomDB customDB = new CustomDB();
-    @Override
-    protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        String title = req.getParameter("deleteId");
-
-        customDB.delete(UUID.fromString(title));
-
-        req.getRequestDispatcher("/home").forward(req, resp);
-    }
-}
