+++ Lesson20-UI-Reaction/src/main/java/config/CustomDB.java
@@ -1,4 +1,6 @@
-import domain.Fridge;
+package config;
+
+import domain.Product;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -6,13 +8,13 @@ import java.util.Optional;
 import java.util.UUID;
 
 public class CustomDB {
-    public static List<Fridge> fridgeList = new ArrayList<>();
+    public static List<Product> fridgeList = new ArrayList<>();
 
-    public void create(Fridge fridge){
+    public void create(Product fridge){
         fridge.setId(UUID.randomUUID());
         fridgeList.add(fridge);
     }
-    public List<Fridge> getByTitle(String title){
+    public List<Product> getByTitle(String title){
         if(title == null || title.isBlank()){
             return fridgeList;
         }
@@ -21,7 +23,7 @@ public class CustomDB {
                 .toList();
     }
     public void delete(UUID uuid){
-        Optional<Fridge> first = fridgeList.stream()
+        Optional<Product> first = fridgeList.stream()
                 .filter(fridge -> fridge.getId().equals(uuid))
                 .findFirst();
