+++ Lesson20-UI-Reaction/src/main/java/servlets/DeleteServlet.java
@@ -1,4 +1,6 @@
-import domain.Fridge;
+package servlets;
+
+import config.CustomDB;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -6,23 +8,16 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
+import java.util.UUID;
 
-@WebServlet("/add")
-public class AddServlet extends HttpServlet {
+@WebServlet("/delete")
+public class DeleteServlet extends HttpServlet {
     private final CustomDB customDB = new CustomDB();
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        String title = req.getParameter("title");
-        float height = Float.parseFloat(req.getParameter("height"));
-        float weight = Float.parseFloat(req.getParameter("weight"));
+        String title = req.getParameter("deleteId");
 
-        Fridge fridge = new Fridge();
-
-        fridge.setTitle(title);
-        fridge.setHeight(height);
-        fridge.setWeight(weight);
-
-        customDB.create(fridge);
+        customDB.delete(UUID.fromString(title));
 
         req.getRequestDispatcher("/home").forward(req, resp);
     }
